# cmd/server/docs/swagger.yaml
definitions:
  database.NumericDataType:
    properties:
      hasAutoIncrement:
        type: boolean
      hasDigit:
        type: boolean
      hasSize:
        type: boolean
      type:
        type: string
    type: object
  database.StringDataType:
    properties:
      hasSize:
        type: boolean
      hasValues:
        type: boolean
      type:
        type: string
    type: object
  models.ColMetaData:
    properties:
      colName:
        type: string
      hasAutoIncrement:
        type: boolean
      hasDefault: {}
      hasDigit:
        type: boolean
      hasSize:
        type: boolean
      hasValues:
        type: boolean
      isUnique:
        type: boolean
      type:
        type: string
    type: object
  models.ColValues:
    properties:
      autoIncrement:
        type: boolean
      colName:
        type: string
      default: {}
      isNull:
        type: boolean
      isPk:
        type: boolean
      isUnique:
        type: boolean
      size:
        type: integer
      type:
        type: string
      value: {}
    type: object
  models.History:
    properties:
      id:
        type: integer
      message:
        type: string
      time:
        type: string
    type: object
  models.InsertRowRequest:
    properties:
      data:
        items:
          $ref: '#/definitions/models.RowItem'
        type: array
      tableName:
        type: string
    type: object
  models.RowItem:
    properties:
      columnName:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  router.DeleteTableRequest:
    properties:
      tableName:
        type: string
      verificationQuery:
        type: string
    type: object
  router.ListRowsResponse:
    properties:
      activeTable:
        type: string
      cols:
        items:
          $ref: '#/definitions/models.ColMetaData'
        type: array
      hasNextPage:
        type: boolean
      page:
        type: integer
      rowCount:
        type: integer
      rows:
        items: {}
        type: array
      totalPages:
        type: integer
    type: object
  service.FormDatatype:
    properties:
      numericType:
        items:
          $ref: '#/definitions/database.NumericDataType'
        type: array
      stringType:
        items:
          $ref: '#/definitions/database.StringDataType'
        type: array
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/history:
    get:
      consumes:
      - application/json
      description: get paginated list of query history
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.History'
                  type: array
              type: object
      summary: List query history
      tags:
      - history
  /api/v1/history/recent:
    get:
      consumes:
      - application/json
      description: get the last 10 query history entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.History'
                  type: array
              type: object
      summary: List recent history
      tags:
      - history
  /api/v1/tables:
    delete:
      consumes:
      - application/json
      description: drop a table from the database
      parameters:
      - description: Delete table request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.DeleteTableRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a table
      tags:
      - tables
    get:
      consumes:
      - application/json
      description: get list of all tables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: {}
            type: object
      summary: list of all tables
      tags:
      - tables
  /api/v1/tables/{tableName}:
    get:
      consumes:
      - application/json
      description: get paginated list of rows for a specific table
      parameters:
      - description: Table Name
        in: path
        name: tableName
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - description: Column to order by
        in: query
        name: column
        type: string
      - description: Order direction (ASC/DESC)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/router.ListRowsResponse'
              type: object
      summary: List rows of a table
      tags:
      - rows
  /api/v1/tables/{tableName}/columns:
    get:
      consumes:
      - application/json
      description: get list of all columns for a specific table
      parameters:
      - description: Table Name
        in: path
        name: tableName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ColMetaData'
                  type: array
              type: object
      summary: List columns of a table
      tags:
      - columns
  /api/v1/tables/{tableName}/form:
    get:
      consumes:
      - application/json
      description: get metadata for columns to build an insert or update form
      parameters:
      - description: Table Name
        in: path
        name: tableName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    Cols:
                      items:
                        $ref: '#/definitions/models.ColMetaData'
                      type: array
                  type: object
              type: object
      summary: Get row insert/update form metadata
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: insert a new row or update an existing row if hash is provided
      parameters:
      - description: Table Name
        in: path
        name: tableName
        required: true
        type: string
      - description: Row hash (for update)
        in: query
        name: hash
        type: string
      - default: 1
        description: Page number (for update context)
        in: query
        name: page
        type: integer
      - description: Row data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.InsertRowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      summary: Insert or update a row
      tags:
      - rows
  /api/v1/tables/{tableName}/row/{hash}:
    delete:
      consumes:
      - application/json
      description: delete a specific row by its hash
      parameters:
      - description: Table Name
        in: path
        name: tableName
        required: true
        type: string
      - description: Row hash
        in: path
        name: hash
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete a row
      tags:
      - rows
  /api/v1/tables/form/new:
    get:
      consumes:
      - application/json
      description: get available numeric and string data types for creating a new
        table
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FormDatatype'
              type: object
      summary: Get data types for new table form
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: create a new table with specified name and columns
      parameters:
      - description: Table Schema
        in: body
        name: body
        required: true
        schema:
          properties:
            inputs:
              items:
                $ref: '#/definitions/models.ColValues'
              type: array
            tableName:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new table
      tags:
      - tables
swagger: "2.0"